modern C ?

+++ libft/srcs/str/ft_atoi.c
@@ -6,17 +6,16 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/26 11:11:40 by tmaluh            #+#    #+#             */
-/*   Updated: 2018/10/26 19:26:47 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/04/04 01:23:00 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/lft_str.h"
-#include "../../includes/lft_ctype.h"
+#include "libft.h"
 
-int					ft_atoi(string str)
+int32_t	ft_atoi(string str)
 {
-	int		num;
-	int		sign;
+	int32_t	num;
+	int8_t	sign;
 
 	sign = 0;
 	num = 0;
